requirements.txt
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
attrs==23.1.0
bcrypt==4.0.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==2.2.1
contourpy==1.1.0
cycler==0.11.0
ecdsa==0.18.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.100.1
filelock==3.12.4
fonttools==4.42.1
frozenlist==1.4.0
h11==0.14.0
httptools==0.6.0
idna==3.4
importlib-resources==6.0.1
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
matplotlib==3.7.2
msgpack==1.0.7
numpy==1.25.2
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
passlib==1.7.4
Pillow==10.0.0
protobuf==4.24.4
psycopg2-binary==2.9.7
pyasn1==0.5.0
pydantic==2.3.0
pydantic_core==2.6.3
pyflame==0.3.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
ray==2.7.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
snakeviz==2.2.0
sniffio==1.3.0
starlette==0.27.0
threadpoolctl==3.2.0
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
zipp==3.16.2
